GiftCardDelegate: Convert CardEncrypter from object to normal class
This will make it easier to test classes that are using CardEncrypter
